Changed up how the code works and added color.
Instead of having multiple if statements etc I changed it to have a
list of APIs which it will check along with the URL and PoC if it is
vulnerable.

These changes have been kind of untested so there may be errors/false positives.

Also added color to the text using colorama to easily differentiate whether
it was vulnerable to an API or not.